Allow multiple base controller classes to be set
This changes the `base_controller_class` configuration option to also
accept an array of class names, so that the custom payload is appended
in multiple controller base classes.

Why? Because some applications have controllers that inherit from
multiple base classes. Example: `ActionController::API` and
`ActionController::Base`. If you have an an application with an API and
also an admin panel, that would certainly be the case.

Until now, using a custom payload was only possible in the controllers
that inherit from the configured `base_controller_class`. With this
change, the custom payload feature can be used in all controllers -- if
their base classes are set.