Display halted callbacks in log for Application Controller
Fix https://github.com/roidrage/lograge/issues/258

Sometimes when you do the final render or redirection, it can be halted by a controller callback. With the existing code, it can very hard to understand where it stopped.

With this commit, the intent is to provide the method name where it stops.